feat: Create Dockerfile and publishing workflows
This patch adds the Dockerfile for the `contane/yamllint` Docker image.
Dependencies are kept up-to-date via Renovate. release-please is used to
create release proposal PRs when pushing/merging to main. Once a
proposal is itself merged, release-please creates a GitHub release.
Then, the publish workflow builds and pushes a tagged image to Docker
Hub.