chore: Publish README to Docker Hub
This ensures that the project description on Docker Hub is always synced
with README.md when a release is made.

README.md is updated to include links to GitHub and Docker Hub such that
people viewing the project on either platform know where to go for the
source code and/or image.